refactor: relocate validation logic
