+++ src/validation/plugin/rules/manifest-category.ts
