+++ src/validation/plugin/rules/manifest-schema.ts
