+++ src/validation/plugin/rules/manifest-urls-exist.ts
